+++ projects/ngx-validation-messages/src/lib/components/ngx-mat-validation-messages/ngx-mat-validation-message-item.component.ts
@@ -1,0 +1,14 @@
+import { Component, Input } from '@angular/core';
+
+/**
+ * Internal helper component for attribute-mode ngxValidationMessages.
+ */
+@Component({
+  standalone: true,
+  selector: 'mat-error',
+  template: '{{ message }}',
+})
+export class NgxMatValidationMessageItemComponent {
+  @Input()
+  public message = '';
+}